[Skyrim] WARN about CollegeOfWinterholdImmersive.esp/EnhancedLightsandFX.esp incompatibility.
From CollegeOfWinterholdImmersive.esp's readme:
"Enhanced lights & FX will make changes to some of the same references as
this mod as well as adding it's own to light sources which may have been
[re]moved, therefore this cannot be overcome by manipulating the load order.
To overcome this incompatibility, open ELnFX in Tes5Edit (There are TONS of
tutorials on it, but its fairly intuitive anyway) and remove any of it's
edits to the College cells (WinterholdCollege[somethingorother]).
This will keep the rest of ELnFX's changes whilst preventing conflict."

I can't remember if the ELFX - *.esp's are adding onto the main module
or replace it, so I've set the WARN to trigger if any of them are found.


git-svn-id: http://better-oblivion-sorting-software.googlecode.com/svn@7200 bfee223e-1fc4-11de-b650-e715bd6d7cf1